android kit refactor

+++ kits/android-kit.sh
@@ -3,33 +3,10 @@
 WORKSPACE="$DEVELOPMENT_KIT_SDK_HOME/Android"
 mkdir -p $WORKSPACE
 
-# Set the mode for this script (install or remove)
-MODE="install"
+MODE=$(curl -o- "https://raw.githubusercontent.com/FirzenYogesh/development-kit/main/commons/task-mode.sh" | bash -s $1)
+OS=$(curl -o- "https://raw.githubusercontent.com/FirzenYogesh/development-kit/main/commons/get-os.sh" | bash)
 
-if [[ -z "$1" ]]; then
-    MODE="install"
-else
-    if [[ $1 == "uninstall" ]] || [[ $1 == "remove" ]] || [[ $1 == "purge" ]]; then
-        MODE="uninstall"
-    fi
-fi
-
-if [[ $MODE == "install" ]]; then
-    cd $WORKSPACE
-    if [[ "$OSTYPE" == "linux"* ]]; then
-        wget -O cmd-tools.zip https://dl.google.com/android/repository/commandlinetools-linux-6609375_latest.zip
-    elif [[ "$OSTYPE" == "darwin"* ]]; then
-        wget -O cmd-tools.zip https://dl.google.com/android/repository/commandlinetools-mac-6609375_latest.zip
-    else
-        wget -O cmd-tools.zip https://dl.google.com/android/repository/commandlinetools-win-6609375_latest.zip
-    fi
-
-    if [[ -d tools ]]; then
-        rm -rf tools
-    fi
-    unzip cmd-tools.zip
-    rm cmd-tools.zip
-
+setEnv() {
     if [[ -z "$ANDROID_HOME" ]]; then
         # Setting up env
         echo 'export ANDROID_HOME="$DEVELOPMENT_KIT_SDK_HOME/Android"' >> $DEVELOPMENT_KIT_ENV
@@ -41,6 +18,25 @@ if [[ $MODE == "install" ]]; then
         echo 'export PATH="$ANDROID_HOME/platform-tools:$PATH"' >> $DEVELOPMENT_KIT_PATHS
         echo 'export PATH="$ANDROID_SDK_ROOT:$PATH"' >> $DEVELOPMENT_KIT_PATHS
     fi
+}
+
+if [[ $MODE == "install" ]]; then
+    cd $WORKSPACE
+    if [[ "$OS" == "linux"* ]]; then
+        wget -O cmd-tools.zip https://dl.google.com/android/repository/commandlinetools-linux-6609375_latest.zip
+    elif [[ "$OS" == "darwin"* ]]; then
+        wget -O cmd-tools.zip https://dl.google.com/android/repository/commandlinetools-mac-6609375_latest.zip
+    else
+        wget -O cmd-tools.zip https://dl.google.com/android/repository/commandlinetools-win-6609375_latest.zip
+    fi
+
+    if [[ -d tools ]]; then
+        rm -rf tools
+    fi
+    unzip cmd-tools.zip
+    rm cmd-tools.zip
+
+    setEnv
 
     source "$DEVELOPMENT_KIT_MAIN"
 
@@ -61,7 +57,9 @@ if [[ $MODE == "install" ]]; then
     yes | sdkmanager "extras;google;instantapps"
     yes | sdkmanager "extras;google;market_licensing"
     yes | sdkmanager --licenses
-else
+elif [[ $MODE == "fix-env" ]]; then
+    setEnv
+elif [[ $MODE == "uninstall" ]]; then
     if [[ -d "$ANDROID_HOME" ]]; then
         rm -rf $ANDROID_HOME
         echo "Please remove the following lines in $DEVELOPMENT_KIT_ENV"
@@ -76,4 +74,7 @@ else
     else
         echo "Android SDK is not installed"
     fi
+else
+    echo "Unsupported Operation"
+    exit 1
 fi
